First Check In
So, I got this to run on mobile to begin with.  For some reason the movement isn't working.  Instead of returning the value, it's returning nothing.  It's really pissing me off.  I even made the value static in case it was somehow creating two classes and the class being called wasn't the class making the changes.  Gave it like an hour, and I just can't keep going.  I'll have to just continue and figure it out later hopefully.