Made boss harder
SO, fun story.  You can submit a name, but it doesn't actually let you change what's in the field right now.  Maybe when I make this work as an actual android project again it will work the way it's supposed to.  I just had so many problems getting this to work at all that I just had it running as a computer app.  Sad.